Fixed polyak_avg being 0.999 instead of 0.001

+++ src/critics/td3_critic.py
@@ -104,7 +104,7 @@ class Td3Critic(nn.Module):
         target_qf2_values = self.qf2_target(target_qf_inputs)
         target_q_values = torch.min(target_qf1_values, target_qf2_values)
         
-        target = train_data['next']['reward'] + self.gamma * target_q_values * (1 - train_data['next']['done'].type(torch.uint8))
+        target = train_data['next']['reward'] + self.gamma * target_q_values * (1 - train_data['next']['done'].type(torch.float))
         target = target.detach()
         
         return target
